rf_center becomes rf_reference.
It turns out the the name "rf_center"  is not
an accurate name for this data.  It is sometimes
the RF center as provided by the observatory
metadata and at other times it is the value of
an LO (perhaps one of a series of LOs).

Thus it is more accurate to call this an RF reference.